pyproject.toml
[tool.poetry]
name = "django_template"
version = "0.1.1"
description = "django template"
authors = ["L.Lu <luliangce@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-jose = "^3.1.0"
django = "^3.1.12"
django-ninja = "^0.16.1"
django-cors-headers = "^3.10.0"

[tool.poetry.dev-dependencies]
yapf = "^0.29.0"
pylint = "^2.4.4"
rope = "^0.16.0"
ipython = "^7.13.0"
[[tool.poetry.source]]
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
name = "ts"
default = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
